requirements.txt
Django==2.1.5
django-filebrowser-no-grappelli==3.7.8
django-tinymce4-lite==1.7.4
idna==2.8
lxml==4.3.2
Pillow==6.1.0
urllib3==1.24.1
websocket-client==0.56.0
requests==2.21.0
gunicorn==19.9.0